prom_metrics_cli/requirements.txt
matplotlib==3.6.2
pandas_ods_reader==0.1.4
seaborn==0.12.1
datetime-matcher==0.2.1
